layers/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile layers/requirements.in
#
aws-lambda-powertools==3.5.0
    # via -r layers/requirements.in
boto3==1.36.16
    # via -r layers/requirements.in
botocore==1.36.16
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.1
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
google-auth==2.38.0
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.1
    # via gspread
gspread==6.1.4
    # via -r layers/requirements.in
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   aws-lambda-powertools
    #   boto3
    #   botocore
mypy-boto3-dynamodb==1.36.0
    # via -r layers/requirements.in
mypy-boto3-s3==1.36.15
    # via -r layers/requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
python-dateutil==2.9.0.post0
    # via botocore
pytz==2025.1
    # via -r layers/requirements.in
requests==2.32.3
    # via
    #   -r layers/requirements.in
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
s3transfer==0.11.2
    # via boto3
six==1.17.0
    # via python-dateutil
typing-extensions==4.12.2
    # via
    #   aws-lambda-powertools
    #   mypy-boto3-dynamodb
    #   mypy-boto3-s3
urllib3==1.26.20
    # via
    #   botocore
    #   requests
